[CICD] Minor changes to fix cicd
