+++ samples/views/sample.py
@@ -1180,10 +1180,6 @@ class SampleRenameForm(forms.Form):
 
     def clean_new_name(self):
         new_name = self.cleaned_data["new_name"]
-        name_format, match = sample_names.sample_name_format(new_name, with_match_object=True)
-        if name_format is None:
-            raise ValidationError(_("This sample name is not valid."), code="invalid")
-        utils.check_sample_name(match, self.user)
         if sample_names.does_sample_exist(new_name):
             raise ValidationError(_("This sample name exists already."), code="duplicate")
         return new_name
@@ -1196,19 +1192,6 @@ class SampleRenameForm(forms.Form):
             if new_name == old_name:
                 self.add_error("new_name", ValidationError(_("The new name must be different from the old name."),
                                                            code="invalid"))
-            old_name_format = sample_names.sample_name_format(old_name)
-            possible_new_name_formats = settings.SAMPLE_NAME_FORMATS[old_name_format].get("possible_renames", set()) \
-                if old_name_format else set()
-            name_format = sample_names.sample_name_format(new_name)
-            if name_format not in possible_new_name_formats:
-                error_message = ngettext("New name must be a valid “%(sample_formats)s” name.",
-                                          "New name must be a valid name of one of these types: %(sample_formats)s.",
-                                          len(possible_new_name_formats))
-                self.add_error("new_name", ValidationError(
-                    error_message,
-                    params={"sample_formats": format_enumeration(
-                        sample_names.verbose_sample_name_format(name_format) for name_format in possible_new_name_formats)},
-                    code="invalid"))
 
         return cleaned_data
