+++ static/js/datatables_lang.js
@@ -1,0 +1,19 @@
+function makeDatatablesLang(extra) {
+    var dict = {
+        processing:     gettext("Processing..."),
+        search:         gettext("Search :"),
+        lengthMenu:     gettext("Show _MENU_ entries"),
+        info:           gettext("Showing _START_ to _END_ of total _TOTAL_ entries"),
+        infoEmpty:      gettext("Showing 0 to 0 of total 0 entries"),
+        infoFiltered:   gettext("(filtered from total _MAX_ entries)"),
+        infoPostFix:    "",
+        loadingRecords: gettext("Loading records..."),
+        zeroRecords:    gettext("No entries found"),
+        emptyTable:     gettext("No data available"),
+        aria: {
+            sortAscending:  gettext(": activate to sort column ascending"),
+            sortDescending: gettext(": activate to sort column descending")
+        }
+    };
+    return Object.assign(dict, extra);
+}
