+++ static/js/lab_notebook.js
@@ -22,23 +22,7 @@ $(document).ready(function () {
     ];
     $(manualTables.join(', ')).addClass('manual-datatable');
     
-    // Define langDict for DataTables (copied from juliabase)
-    var langDict = {
-        processing:     gettext("Processing..."),
-        search:         gettext("Search&nbsp;:"),
-        lengthMenu:     gettext("Show _MENU_ entries"),
-        info:           gettext("Showing _START_ to _END_ of total _TOTAL_ entries"),
-        infoEmpty:      gettext("Showing 0 to 0 of total 0 entries"),
-        infoFiltered:   gettext("(filtered from total _MAX_ entries)"),
-        infoPostFix:    "",
-        loadingRecords: gettext("Loading records..."),
-        zeroRecords:    gettext("No entries found"),
-        emptyTable:     gettext("No data available"),
-        aria: {
-            sortAscending:  gettext(": activate to sort column ascending"),
-            sortDescending: gettext(": activate to sort column descending")
-        }
-    };
+    var langDict = makeDatatablesLang();
 
     // Specific initializations moved from juliabase
     $('#experiment').DataTable({
